requirements.txt
openai>=1.60.0
vercel-blob==0.4.2
numpy>=1.26.0
python-dotenv==1.0.1
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
requests>=2.31.0
authlib==1.3.2
Pillow>=10.0.0
pytesseract==0.3.13
python-docx==1.1.2
psycopg2-binary>=2.9.9